requirements.txt
RPi.GPIO==0.7.1
watchdog==3.0.0
